tests: moving serialization tests to their own folder and namespace
