+++ Assets/Tests/Editor/Serialization.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f4271f993f5be45bdbebcc018c455da7
+guid: 2bffe0a4629483b41b8c48f97ff3dc30
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
