+++ Assets/Tests/Runtime/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d42e7014f0ce5084bb2f8b11404d6774
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
