+++ Assets/Tests/Runtime/Serialization/ArrayWriterTest.cs
@@ -1,6 +1,7 @@
 using Mirage.Serialization;
 using NUnit.Framework;
-namespace Mirage.Tests.Runtime
+
+namespace Mirage.Tests.Runtime.Serialization
 {
     [TestFixture]
     public class ArrayWriterTest
