+++ Assets/Tests/Runtime/Serialization/MessageInheritanceTest.cs.meta
