+++ Assets/Tests/Runtime/Serialization/MessageTests.cs
@@ -2,7 +2,7 @@ using System;
 using Mirage.Serialization;
 using NUnit.Framework;
 
-namespace Mirage.Tests.Runtime
+namespace Mirage.Tests.Runtime.Serialization
 {
     public class AssertionMethodAttribute : Attribute { }
