+++ Assets/Tests/Runtime/Serialization/NetworkPlayerTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a1a87ad2c7e74dc69138ba36f583640
+guid: 10d85f8072533d14e809066d696cc9a0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
