+++ Assets/Tests/Runtime/Serialization/NetworkReaderPoolTest.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Mirage.Serialization;
 using NUnit.Framework;
 
-namespace Mirage
+namespace Mirage.Tests.Runtime.Serialization
 {
     public class NetworkReaderPoolTest
     {
