+++ Assets/Tests/Runtime/Serialization/NetworkWriterPoolTest.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using Mirage.Serialization;
 using NUnit.Framework;
 
-namespace Mirage.Tests.Runtime
+namespace Mirage.Tests.Runtime.Serialization
 {
     public class NetworkWriterPoolTest
     {
