+++ Assets/Tests/Runtime/Serialization/NetworkWriterTest.cs
@@ -4,7 +4,7 @@ using Mirage.Serialization;
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Mirage.Tests.Runtime
+namespace Mirage.Tests.Runtime.Serialization
 {
     [TestFixture]
     public class NetworkWriterTest
