+++ Assets/Tests/Runtime/Serialization/NetworkWriterTest.cs.meta
