+++ Assets/Tests/Runtime/Serialization/ScriptableObjectWriterTest.cs
@@ -2,7 +2,7 @@ using Mirage.Serialization;
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Mirage.Tests.Runtime
+namespace Mirage.Tests.Runtime.Serialization
 {
     internal class MyScriptableObject : ScriptableObject
     {
