+++ Assets/Tests/Runtime/Serialization/ScriptableObjectWriterTest.cs.meta
