+++ Assets/Tests/Runtime/SyncVarTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a5e2205a841a1463fa02a35274d6d76a
+guid: ebd1e02f61311a442bded5cf8cb0ac19
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
