+++ Assets/Tests/Runtime/SyncVarVirtualTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c3e3ffcfc0a8d764e8a2cdb0473afab9
+guid: ce8496b18e8c4784b8561bbc8ef22253
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
